refactor: Google Cloudに関連するファイルを切り分ける
